+++ example/cpp/classifier.h
@@ -1,5 +1,16 @@
 // C++17: copy this small wrapper into your application and link vision_inference.
 #pragma once
+#if defined(_MSC_VER)
+#if _MSC_VER < 1916
+#error Use Visual Studio 2017 15.9 (v141 14.16) or newer.
+#endif
+#if !defined(_MSVC_LANG) || _MSVC_LANG < 201703L
+#error Enable C++17: Project Properties > C/C++ > Language > /std:c++17.
+#endif
+#if !defined(__cpp_noexcept_function_type)
+#error Enable /Zc:noexceptTypes and remove /Zc:noexceptTypes- before including ONNX Runtime.
+#endif
+#endif
 #include "vision_inference.h"
 #include <opencv2/imgcodecs.hpp>
 #include <chrono>